third_party/blink/web_tests/fullscreen/video-fail-to-enter-full-screen.html

<!DOCTYPE html>
<p>This tests that the video should not be rendered in full screen if requestFullScreen() fails.</p>
<video id="video" width="300"></video>
<script src="full-screen-test.js"></script>
<script src="../media/media-controls.js"></script>
<script>
    var video = document.getElementById('video');
    video.webkitRequestFullScreen();
    setTimeout(function() {
        endTest();
    }, 300);
</script>
