third_party/blink/web_tests/fullscreen/video-fail-to-enter-full-screen-expected.txt

CONSOLE WARNING: Failed to execute 'requestFullscreen' on 'Element': API can only be initiated by a user gesture.
This tests that the video should not be rendered in full screen if requestFullScreen() fails.

END OF TEST
